requirements.txt
isodate==0.6.1
python-dotenv==1.0.0
duffel-api==0.6.0
pandas==2.0.3
jupyter==1.0.0